feat: Use OpenFolderDialog in .NET 8 WPF
